ECON482: Biology, Neuroscience and Economics
This paper aims to leave aside ethical and equity concerns, rather focus on discussing potential issues and consequences of this oversampling effect in the economics profession using economics language. As identities of modern humans are intersectional, leading to materials for our reasoning to include both baseline and intersectional tradeoffs, the oversampling of economists’ few identities clusters results in hidden serial correlations with bias in error variances not accounted for, when findings got scaled up for public policy.